python-bareos: handling of the TLS protocol version parameter
Allow to specify TLS protocol version also for FileDaemon connections.
Before it could only be specified for Director connections.
