+++ python-bareos/bareos/bsock/__init__.py
@@ -22,9 +22,12 @@
 from bareos.exceptions import *
 from bareos.util.password import Password
 from bareos.bsock.connectiontype import ConnectionType
+from bareos.bsock.constants import Constants
+from bareos.bsock.filedaemon import FileDaemon
 from bareos.bsock.directorconsole import DirectorConsole
 from bareos.bsock.directorconsolejson import DirectorConsoleJson
-from bareos.bsock.filedaemon import FileDaemon
+from bareos.bsock.protocolversions import ProtocolVersions
+from bareos.bsock.tlsversionparser import TlsVersionParser
 
 # compat
 from bareos.bsock.bsock import BSock
